Refactoring, eureka and config initial setup
