+++ pi-accounting/pom.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0"
-		 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
 
     <parent>
         <groupId>ru.gb.dev.spring</groupId>
@@ -10,12 +10,12 @@
         <artifactId>pi</artifactId>
     </parent>
 
-	<artifactId>pi-accounting</artifactId>
+    <artifactId>pi-accounting</artifactId>
     <version>1.0.0</version>
-	<name>${project.artifactId}</name>
-	<description>Accounting service</description>
+    <name>${project.artifactId}</name>
+    <description>Accounting service</description>
 
-	<dependencies>
+    <dependencies>
         <dependency>
             <groupId>org.springframework.cloud</groupId>
             <artifactId>spring-cloud-starter-config</artifactId>
