+++ pi-accounting/src/main/java/ru/gb/dev/spring/piaccounting/PiAccountingApplication.java
@@ -1,4 +1,4 @@
-package ru.gb.dev.spring.piappweb;
+package ru.gb.dev.spring.piaccounting;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -8,11 +8,11 @@ import org.springframework.cloud.openfeign.EnableFeignClients;
 @EnableFeignClients
 @EnableDiscoveryClient
 @SpringBootApplication
-public class PiAppWebApplication {
+public class PiAccountingApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(PiAppWebApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(PiAccountingApplication.class, args);
+    }
 
 }
