+++ pi-accounting/src/main/resources/application.yml
@@ -1,0 +1,3 @@
+spring:
+  application:
+    name: accounting
