+++ pi-accounting/src/test/java/ru/gb/dev/spring/piaccounting/PiAccountingApplicationTests.java
@@ -1,4 +1,4 @@
-package ru.gb.dev.spring.piappweb;
+package ru.gb.dev.spring.piaccounting;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -7,11 +7,11 @@ import org.springframework.test.context.junit4.SpringRunner;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
-public class PiAppWebApplicationTests {
+public class PiAccountingApplicationTests {
 
-	@Test
-	public void contextLoads() {
-	}
+    @Test
+    public void contextLoads() {
+    }
 
 }
