+++ pi-server-config/src/main/java/ru/gd/dev/spring/pi/server/config/PiServerConfigApplication.java
@@ -5,14 +5,13 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.config.server.EnableConfigServer;
 
 /**
- *
  * http://127.0.0.1:8888/pi-server-discovery.properties
  * http://127.0.0.1:8888/pi-app-api.properties
  * http://127.0.0.1:8888/pi-app-web.properties
- *
  */
 
 @EnableConfigServer
+
 @SpringBootApplication
 public class PiServerConfigApplication {
