+++ pi-server-config/src/test/java/ru/gd/dev/spring/pi/server/config/PiServerConfigApplicationTests.java
@@ -9,9 +9,9 @@ import org.springframework.test.context.junit4.SpringRunner;
 @SpringBootTest
 public class PiServerConfigApplicationTests {
 
-	@Test
-	public void contextLoads() {
-	}
+    @Test
+    public void contextLoads() {
+    }
 
 }
