+++ pi-server-discovery/src/main/java/ru/gb/dev/spring/pi/server/discovery/PiServerDiscoveryApplication.java
@@ -8,8 +8,8 @@ import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
 @SpringBootApplication
 public class PiServerDiscoveryApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(PiServerDiscoveryApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(PiServerDiscoveryApplication.class, args);
+    }
 
 }
