+++ pi-server-discovery/src/main/resources/application.yml
@@ -2,7 +2,7 @@ server:
   port: 8761
 spring:
   application:
-    name: pi-server-discovery
+    name: eureka
   cloud:
     config:
       fail-fast: true
