+++ pi-server-discovery/src/test/java/ru/gb/dev/spring/pi/server/discovery/PiServerDiscoveryApplicationTests.java
@@ -9,9 +9,9 @@ import org.springframework.test.context.junit4.SpringRunner;
 @SpringBootTest
 public class PiServerDiscoveryApplicationTests {
 
-	@Test
-	public void contextLoads() {
-	}
+    @Test
+    public void contextLoads() {
+    }
 
 }
