+++ pi-server-gateway/src/main/resources/application.yml
@@ -1,0 +1,6 @@
+spring:
+  application:
+    name: gateway
+cloud:
+  config:
+    fail-fast: true
