+++ pi-server-gateway/src/test/java/ru/gb/dev/spring/pi/server/gateway/PiServerGatewayApplicationTests.java
@@ -1,4 +1,4 @@
-package ru.gb.dev.spring.pi.server.proxy;
+package ru.gb.dev.spring.pi.server.gateway;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -7,11 +7,11 @@ import org.springframework.test.context.junit4.SpringRunner;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest
-public class PiServerProxyApplicationTests {
+public class PiServerGatewayApplicationTests {
 
-	@Test
-	public void contextLoads() {
-	}
+    @Test
+    public void contextLoads() {
+    }
 
 }
